CYS - reduce editor instance re-render
